refactor: move getUpdData to utils; use a shallow copy if newChats as prevChats
